Bump MSRV to 1.68
syn 2.0.107 raised its MSRV to 1.68, which caused our MSRV CI to break.
We depend on syn in the cairo-example, xtrace-example and some dev
dependencies (via criterion). Nothing of this is important for x11rb, so
normally I would just pin the previous syn version in MSRV CI. However,
we also have syn 1 in the dependency tree of cairo-example, so "the
usual approach" does not work cleanly.

Instead, let's follow this MSRV bump and bump our own MSRV.

This commit is just a search-and-replace operation. Cleanups will come
in later commits.

Signed-off-by: Uli Schlachter <psychon@znc.in>