Update to raw-window-handle 0.6
There has been a breaking API change in raw-window-handle. The traits
HasRawDisplayHandle and HasRawWindowHandle have been deprecated in favor
of the new HasDisplayHandle and HasWindowHandle traits. Instead of a raw
handle, a handle containing a lifetime is handed out via these traits.
Internally, we still use the raw API and convert to the new handles.

Signed-off-by: Uli Schlachter <psychon@znc.in>